Add parent component to validate CheckboxGroup
