+++ frontend/src/components/CheckboxGroup/CheckboxGroup.tsx
@@ -1,15 +1,12 @@
-import { List } from 'immutable';
 import React, { FunctionComponent, useEffect, useState } from 'react';
 import { Checkbox, CheckboxProps, DropdownItemProps, Form, Segment } from 'semantic-ui-react';
 import styled from 'styled-components';
-import { Filters, OperationStepMap } from '../../types/operations';
-import { BasicModal } from '../BasicModal';
 
 interface ComponentProps {
-  steps?: List<OperationStepMap>;
   options: DropdownItemProps[];
   selectedOptions?: string[];
   onUpdateOptions?: (options: string) => void;
+  validateCheckboxOptions?: (option: string, checkboxState: boolean | undefined) => void;
 }
 
 const StyledSegment = styled(Segment)`
@@ -21,8 +18,6 @@ const CheckboxGroup: FunctionComponent<ComponentProps> = (props) => {
   const [checkboxes, addCheckboxes] = useState<string[] | undefined>(
     props.selectedOptions && props.selectedOptions.length > 0 ? props.selectedOptions : [],
   );
-  const [showModal, setShowModal] = useState<boolean>(false);
-  const [modalMessage, setModalMessage] = useState<string>('');
 
   useEffect(() => {
     if (props.onUpdateOptions) {
@@ -38,201 +33,14 @@ const CheckboxGroup: FunctionComponent<ComponentProps> = (props) => {
       ? checkboxes?.concat(data.value as string)
       : checkboxes?.filter((checkbox) => checkbox !== data.value);
     addCheckboxes(updatedCheckboxes);
-    validateCheckboxOptions(data.value as string, data.checked);
+    props.validateCheckboxOptions &&
+      props.validateCheckboxOptions(data.value as string, data.checked);
   };
 
   const isChecked = (value: string): boolean => {
     return props.selectedOptions && props.selectedOptions.length > 0
       ? !!props.selectedOptions.find((c: string) => c === value)
       : false;
-  };
-
-  const validateCheckboxOptions = (
-    checkboxValue: string,
-    checkboxState: boolean | undefined,
-  ): void => {
-    const { steps } = props;
-    steps &&
-      steps.map((step) => {
-        const options = step.get('query_kwargs') as string;
-        console.log(`Options are ${options}`);
-        toggleModalState(options, checkboxState, checkboxValue);
-      });
-  };
-
-  const toggleModalState = (
-    options: string,
-    checkboxState: boolean | undefined,
-    checkboxValue: string,
-  ): void => {
-    const parsedOptions = parseOptions(options);
-    parsedOptions?.values?.flat().reduce(function (result, query) {
-      if (
-        query &&
-        query.field === checkboxValue &&
-        !checkboxState &&
-        parsedOptions.id === 'filters'
-      ) {
-        setShowModal(true);
-        setModalMessage(
-          `Notice that ${query.field
-            .split('_')
-            .join(' ')} is used in a filter step, deselecting it here may break this query.`,
-        );
-      }
-      if (
-        query &&
-        query.operational_column === checkboxValue.substr(0, checkboxValue.lastIndexOf('_')) &&
-        !checkboxState &&
-        parsedOptions.id === 'aggregate'
-      ) {
-        setShowModal(true);
-        setModalMessage(
-          `Notice that ${query.operational_column
-            .split('_')
-            .join(' ')} is used in an aggregate step, deselecting it here may break this query.`,
-        );
-      }
-      if (parsedOptions.id === 'join_on') {
-        const keys = Object.keys(query.join_on);
-        for (const iterator of keys) {
-          if (iterator === checkboxValue && !checkboxState) {
-            setShowModal(true);
-            setModalMessage(
-              `Notice that ${iterator
-                .split('_')
-                .join(' ')} is used in a join step, deselecting it here may break this query.`,
-            );
-            break;
-          }
-        }
-      }
-      if (
-        query &&
-        query.operational_column === checkboxValue &&
-        !checkboxState &&
-        parsedOptions.id === 'scalar'
-      ) {
-        setShowModal(true);
-        setModalMessage(
-          `Notice that ${query.operational_column
-            .split('_')
-            .join(
-              ' ',
-            )} is used in a scalar transform step, deselecting it here may break this query.`,
-        );
-      }
-      if (query && query.operational_columns && !checkboxState && parsedOptions.id === 'multi') {
-        for (const iterator of query.operational_columns) {
-          if (iterator === checkboxValue && !checkboxState) {
-            setShowModal(true);
-            setModalMessage(
-              `Notice that ${iterator
-                .split('_')
-                .join(
-                  ' ',
-                )} is used in a multi transform step, deselecting it here may break this query.`,
-            );
-            break;
-          }
-        }
-      }
-      if (query && query.over && !checkboxState && parsedOptions.id === 'window') {
-        for (const iterator of query.columns) {
-          if (iterator === checkboxValue && !checkboxState) {
-            setShowModal(true);
-            setModalMessage(
-              `Notice that ${iterator
-                .split('_')
-                .join(' ')} is used in a window step, deselecting it here may break this query.`,
-            );
-            break;
-          }
-        }
-      }
-
-      return result;
-    }, []);
-  };
-
-  const parseOptions = (options: string) => {
-    const parsedOptions = JSON.parse(options);
-    if (parsedOptions.hasOwnProperty('filters')) {
-      const { filters }: Filters = options ? parsedOptions : { filters: [] };
-
-      return {
-        id: 'filters',
-        values: [filters],
-      };
-    }
-    if (parsedOptions.hasOwnProperty('join_on')) {
-      const joinOptions = options
-        ? JSON.parse(options)
-        : {
-            table_name: '',
-            schema_name: '',
-            join_on: {},
-            columns_x: [],
-            columns_y: [],
-            join_how: '',
-          };
-
-      return {
-        id: 'join_on',
-        values: [joinOptions],
-      };
-    }
-    if (
-      parsedOptions.hasOwnProperty('operational_column') &&
-      !parsedOptions.hasOwnProperty('trans_func_name')
-    ) {
-      const aggregateOptions = options
-        ? JSON.parse(options)
-        : { group_by: [], agg_func_name: '', operational_column: '' };
-
-      return {
-        id: 'aggregate',
-        values: [aggregateOptions],
-      };
-    }
-    if (
-      parsedOptions.hasOwnProperty('operational_column') &&
-      parsedOptions.hasOwnProperty('trans_func_name')
-    ) {
-      const scalarOptions = options
-        ? JSON.parse(options)
-        : { group_by: [], agg_func_name: '', operational_column: '' };
-
-      return {
-        id: 'scalar',
-        values: [scalarOptions],
-      };
-    }
-    if (parsedOptions.hasOwnProperty('operational_columns')) {
-      const multiOptions = options
-        ? JSON.parse(options)
-        : {
-            operational_value: '',
-            trans_func_name: '',
-            operational_column: '',
-            operational_columns: [],
-          };
-
-      return {
-        id: 'multi',
-        values: [multiOptions],
-      };
-    }
-    if (parsedOptions.hasOwnProperty('over')) {
-      const windowOptions = options
-        ? JSON.parse(options)
-        : { window_fn: '', order_by: [], term: '', over: [], columns: [] };
-
-      return {
-        id: 'window',
-        values: [windowOptions],
-      };
-    }
   };
 
   const groupIntoRows = (options: DropdownItemProps[]): DropdownItemProps[][] => {
@@ -251,33 +59,24 @@ const CheckboxGroup: FunctionComponent<ComponentProps> = (props) => {
     return rows;
   };
 
-  const toggleShowModal = () => {
-    setShowModal(false);
-  };
-
   return (
-    <>
-      <BasicModal show={showModal} onHide={toggleShowModal}>
-        {modalMessage}
-      </BasicModal>
-      <StyledSegment>
-        {groupIntoRows(props.options).map((row, index) => (
-          <div key={`${index}`} className="row">
-            {row.map(({ key, text, value }) => (
-              <Form.Field key={key} className="col-md-4">
-                <Checkbox
-                  checked={isChecked(value as string)}
-                  label={text}
-                  value={value as string}
-                  onChange={onChange}
-                  className={'selectColumnCheckbox'}
-                />
-              </Form.Field>
-            ))}
-          </div>
-        ))}
-      </StyledSegment>
-    </>
+    <StyledSegment>
+      {groupIntoRows(props.options).map((row, index) => (
+        <div key={`${index}`} className="row">
+          {row.map(({ key, text, value }) => (
+            <Form.Field key={key} className="col-md-4">
+              <Checkbox
+                checked={isChecked(value as string)}
+                label={text}
+                value={value as string}
+                onChange={onChange}
+                className={'selectColumnCheckbox'}
+              />
+            </Form.Field>
+          ))}
+        </div>
+      ))}
+    </StyledSegment>
   );
 };
