+++ frontend/src/components/CheckboxGroup/index.ts
@@ -1,1 +1,2 @@
 export * from './CheckboxGroup';
+export * from './CheckboxGroupValidator';
